+++ src/shareOclock/proMember/ProMemberDAO.java
@@ -1,4 +1,4 @@
-package shareOclock.proMember;
+package shareOclock.member;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -6,204 +6,102 @@ import java.sql.ResultSet;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.apache.tomcat.dbcp.dbcp2.BasicDataSource;
-
 import configuration.Configuration;
-import shareOclock.file.FilesDAO;
-import shareOclock.member.MemberDTO;
 
+public class Pro_memberDAO {
+	public static Pro_memberDAO instance;
 
-
-public class ProMemberDAO {
-	private static ProMemberDAO instance;
-	public synchronized static ProMemberDAO getInstance() {
+	public synchronized static Pro_memberDAO getInstance() {
 		if (instance == null) {
-			System.out.println("디비가 생성되었다.");
-			instance = new ProMemberDAO();
+			System.out.println("디비 생성");
+			instance = new Pro_memberDAO();
 		}
 		return instance;
 	}
 	Connection getConnection() throws Exception {
 		return Configuration.dbs.getConnection();
 	}
-	public int insert(ProMemberDTO dto) throws Exception{
-		String sql = "insert into tb_project_member values(pm_seq.nextval,'M',?,?,?,?,?)";
-		try(
-				Connection con = this.getConnection();
-				PreparedStatement pstat = con.prepareStatement(sql);
-				){
-			pstat.setString(1, dto.getPm_img());
-			pstat.setString(2, dto.getPm_nickname());
-			pstat.setString(3, dto.getPm_name());
-			pstat.setString(4, dto.getPm_email());
-			pstat.setInt(5, dto.getPro_seq());
+
+	public int insertProjectMember(int pro_seq, String pm_nickname, String pm_check) throws Exception{
+		String sql = "insert into tb_project_member values(pm_seq.nextval,?,?,?)";
+		try(Connection con = this.getConnection();
+			PreparedStatement pstat = con.prepareStatement(sql);){
+			pstat.setInt(1, pro_seq);
+			pstat.setString(2, pm_nickname);
+			pstat.setString(3, pm_check);
 			int result = pstat.executeUpdate();
 			con.commit();
-			return result;
+			return result;			
 		}
 	}
-	public ArrayList<String> getThree(String mb_nickname) throws Exception{
-		String sql = "select mb_img, mb_nickname, mb_name from tb_member where mb_nickname = ?";
-		try(
-				Connection con = this.getConnection();
-				PreparedStatement pstat = con.prepareStatement(sql);
-				){
-			pstat.setString(1, mb_nickname);
-			ArrayList<String> al = new ArrayList<>();
+	public List<Integer> getSeqByNickname(String nickname) throws Exception{
+		String sql = "select pro_seq from tb_project_member where pm_nickname=?";
+		try(Connection con = Configuration.dbs.getConnection();
+			PreparedStatement pstat = con.prepareStatement(sql);){
+			pstat.setString(1, nickname);
+			try(ResultSet rs = pstat.executeQuery();){
+				List<Integer> list = new ArrayList<>();
+				while(rs.next()) {
+					int seq = rs.getInt("pro_seq");
+					list.add(seq);
+				}return list;
+			}
+		}
+	}
+	
+	public int getCountBySeq(int seq) throws Exception{
+		String sql = "select count(*) from tb_project_member where pro_seq=?";
+		try(Connection con = Configuration.dbs.getConnection();
+			PreparedStatement pstat = con.prepareStatement(sql);){
+			pstat.setInt(1, seq);
 			try(ResultSet rs = pstat.executeQuery();){
 				if(rs.next()) {
-					String mb_img = rs.getString("mb_img");
-					String mb_nickname1 = rs.getString("mb_nickname");
-					String mb_name = rs.getString("mb_name");
-					System.out.println(mb_img + mb_nickname1 + mb_name);
-					al.add(mb_img);
-					al.add(mb_nickname1);
-					al.add(mb_name);
-				}
-			}
-			return al;
+					int count = rs.getInt(1);
+					return count;
+				}return -1;			
+			}	
+		}
+	}	
 
+	public boolean isValidMember(int pro_seq, String nickname) throws Exception{
+		String sql = "select *  from tb_project_member where pro_seq=? and pm_nickname=?";
+		try(Connection con = Configuration.dbs.getConnection();
+			PreparedStatement pstat = con.prepareStatement(sql);){
+			pstat.setInt(1, pro_seq);
+			pstat.setString(2, nickname);
+			try(ResultSet rs = pstat.executeQuery();){
+				if(rs.next()) {
+					return true;
+				}return false;
+			}
 		}
 	}
-	public int delete(String pm_nickname) throws Exception{
-		String sql = "delete from tb_project_member where pm_nickname = ?";
-		try(
-				Connection con = this.getConnection();
-				PreparedStatement pstat = con.prepareStatement(sql);
-				){
-			pstat.setString(1, pm_nickname);		
+	
+	public int deleteProjectMember(int pro_seq) throws Exception{
+		String sql = "delete from tb_project_member where pro_seq=?";
+		try(Connection con = Configuration.dbs.getConnection();
+			PreparedStatement pstat = con.prepareStatement(sql);){
+			pstat.setInt(1, pro_seq);
 			int result = pstat.executeUpdate();
 			con.commit();
 			return result;
 		}
 	}
-	public int update(ProMemberDTO dto) throws Exception{
-		String sql = "insert into tb_project_member values(pm_seq.nextval,'M',?,?,?,?,?)";
-		try(
-				Connection con = this.getConnection();
-				PreparedStatement pstat = con.prepareStatement(sql);
-				){
-			pstat.setString(1, dto.getPm_img());
-			pstat.setString(2, dto.getPm_nickname());
-			pstat.setString(3, dto.getPm_name());
-			pstat.setString(4, dto.getPm_email());
-			pstat.setInt(5, dto.getPro_seq());
-			int result = pstat.executeUpdate();
-			con.commit();
-			return result;
-		}
-	}
-	public List<ProMemberDTO> selectPM() throws Exception{
-		String sql = "select * from tb_project_member order by pm_seq, pm_name";
-		try(
-				Connection con = this.getConnection();
-				PreparedStatement pstat = con.prepareStatement(sql);
-				ResultSet rs = pstat.executeQuery();
-				){
-			List<ProMemberDTO> list = new ArrayList<>();
-			while(rs.next()) {
-				int pm_seq = rs.getInt("pm_seq");
-				String pm_check = rs.getString("pm_check");
-				String pm_img = rs.getString("pm_img");
-				String pm_nickname = rs.getString("pm_nickname");
-				String pm_name = rs.getString("pm_name");
-				String pm_email = rs.getString("pm_email");
-				int pro_seq = rs.getInt("pro_seq");
-				ProMemberDTO dto = new ProMemberDTO(pm_seq,pm_check, pm_img, pm_nickname, pm_name, pm_email, pro_seq);
-				list.add(dto);
-			}
-			return list;
-		}
-	}
-	public List<ProMemberDTO> selectByNickname(String pm_nickname) throws Exception{
-		String sql = "select * from tb_project_member where pm_nickname = ?";
-		try(
-				Connection con = this.getConnection();
-				PreparedStatement pstat = con.prepareStatement(sql);
-				){
-			pstat.setString(1, pm_nickname);
+	
+	public boolean isProjectHeader(int pro_seq, String nickname) throws Exception{
+		String sql = "select pm_check from tb_project_member where pro_seq=? and pm_nickname=?";
+		try(Connection con = Configuration.dbs.getConnection();
+			PreparedStatement pstat = con.prepareStatement(sql);){
+			pstat.setInt(1, pro_seq);
+			pstat.setString(2, nickname);
 			try(ResultSet rs = pstat.executeQuery();){
-				List<ProMemberDTO> list = new ArrayList<>();
-				while(rs.next()) {
-					int pm_seq = rs.getInt("pm_seq");
-					String pm_check = rs.getString("pm_check");
-					String pm_img = rs.getString("pm_img");
-					String pm_nickname1 = rs.getString("pm_nickname");
-					String pm_name = rs.getString("pm_name");
-					String pm_email = rs.getString("pm_email");
-					int pro_seq = rs.getInt("pro_seq");
-					ProMemberDTO dto = new ProMemberDTO(pm_seq,pm_check, pm_img, pm_nickname1, pm_name, pm_email, pro_seq);
-					list.add(dto);
-				}
-				return list;
+				if(rs.next()) {
+					String pm_check = rs.getString(1);
+					if(pm_check.contentEquals("H")) {
+						return true;
+					}
+				}return false;
 			}
 		}
 	}
-	public List<MemberDTO> selectMemeber(String pm_nickname) throws Exception{
-		String sql = "select * from tb_member where not mb_nickname in (?)";
-		try(
-				Connection con = this.getConnection();
-				PreparedStatement pstat = con.prepareStatement(sql);
-				){
-			pstat.setString(1, pm_nickname);
-			try(ResultSet rs = pstat.executeQuery();){
-				List<MemberDTO> list = new ArrayList<>();
-				while(rs.next()) {
-					System.out.println("while안");
-					String mb_img = rs.getString("mb_img");
-					String mb_nickname = rs.getString("mb_nickname");
-					String mb_pw = rs.getString("mb_pw");
-					String mb_name = rs.getString("mb_name");	
-					String mb_email = rs.getString("mb_email");	
-					String mb_group = rs.getString("mb_group");
-					String mb_phone = rs.getString("mb_phone");
-					String mb_check = rs.getString("mb_check");
-					MemberDTO dto = new MemberDTO(mb_img, mb_nickname, null, mb_name, mb_email, null, null, "M");
-					list.add(dto);
-				}
-				return list;
-			}
-		}
-	}
-	public List<MemberDTO> updateList() throws Exception{
-		String sql = "SELECT m.mb_img, m.mb_nickname, m.mb_pw, m.mb_name, m.mb_email, m.mb_group, m.mb_phone, m.mb_check FROM tb_member m WHERE not exists (select * from tb_project_member p where m.mb_nickname = p.pm_nickname)";
-		try(
-				Connection con = this.getConnection();
-				PreparedStatement pstat = con.prepareStatement(sql);
-				ResultSet rs = pstat.executeQuery();
-				){
-			List<MemberDTO> list = new ArrayList<>();
-			while(rs.next()) {
-				String mb_img = rs.getString("mb_img");
-				String mb_nickname = rs.getString("mb_nickname");
-				String mb_pw = rs.getString("mb_pw");
-				String mb_name = rs.getString("mb_name");	
-				String mb_email = rs.getString("mb_email");	
-				String mb_group = rs.getString("mb_group");
-				String mb_phone = rs.getString("mb_phone");
-				String mb_check = rs.getString("mb_check");
-				MemberDTO dto = new MemberDTO(mb_img, mb_nickname, null, mb_name, mb_email, null, null, "M");
-				list.add(dto);
-			}
-			return list;
-		}
-	}
-	public List<String> searchNickname(String mb_nickname) throws Exception{
-		String sql = "select mb_nickname from tb_member where mb_nickname like ? ";
-		try(
-				Connection con = this.getConnection();
-				PreparedStatement pstat = con.prepareStatement(sql);
-				){
-			pstat.setString(1, mb_nickname);
-			try(ResultSet rs = pstat.executeQuery();){
-				List<String> list = new ArrayList<>();
-				while(rs.next()) {
-					String mb_nickname1 = rs.getString("mb_nickname");					
-					list.add(mb_nickname1);
-				}
-				return list;
-			}
-		}
-	}
-
 }
